infra/qdrant/test/requirements.txt
langchain-qdrant
langchain-community
langchain-huggingface
python-dotenv
qdrant-client
langchain
langchain-text-splitters